Change app name from "n3v3rs text one" to "Deutsch"
Left the original author's name intact. Credit where credit is due. The reason
for the change is that "n3v3rs text one" is somewhat meaningless when browsing
watch faces, and "Deutsch" seems to be a better title. Even better than
"Uhrzeit als Text" which is the title under which the watch face seems to be
listed in the Pebble store.